Declarative Net Request redirect not working on clicked links https://bugs.webkit.org/show_bug.cgi?id=291325 rdar://142017438
Reviewed by NOBODY (OOPS!).

The issue we were seeing here was that if an extension had a DNR redirect rule to redirect
navigations from site A to site B, when clicking on a link for site A, we would commit the navigation
to site A, fail that load, and set the provisional page for that load to `nil`. I believe setting
this page to `nil` is what cause the load to site B to be cancelled (commenting out that code
caused the redirect) to go through. Simply not setting the provisional page to nil if a redirect
would occur wouldn't work here because there's no way to determine that a redirect would occur in
`WebPageProxy::didFailProvisionalLoadForFrameShared`.

The fix here is to run URLs through the compiled byte code for DNR rules before deciding the policy
for the navigation. In doing this, we won't attempt to load site A, we will consult the DNR rules
first, and if a redirect should occur, we will update the request and decide the policy for site B.

Testing:
- Clicking on a link to site A redirects to site B
- Navigating to site A via a navigation bar redirects to site B
- Long pressing on a link for site A shows a preview for site B

* Source/WebCore/contentextensions/ContentExtension.cpp:
(WebCore::ContentExtensions::ContentExtension::populateTopURLActionCacheIfNeeded const):
Don't populate the action cache if the topURL is empty.
(WebCore::ContentExtensions::ContentExtension::populateFrameURLActionCacheIfNeeded const):
Don't populate the action cache if the frameURL is empty.
* Source/WebCore/loader/PolicyChecker.cpp:
(WebCore::PolicyChecker::checkNavigationPolicy):